+++ contracts/minter/src/msg.rs
@@ -7,7 +7,7 @@ use sg721::msg::InstantiateMsg as Sg721InstantiateMsg;
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InstantiateMsg {
     pub base_token_uri: String,
-    pub num_tokens: u64,
+    pub num_tokens: u32,
     pub sg721_code_id: u64,
     pub sg721_instantiate_msg: Sg721InstantiateMsg,
     pub start_time: Timestamp,
@@ -24,7 +24,7 @@ pub enum ExecuteMsg {
     UpdateStartTime(Timestamp),
     UpdatePerAddressLimit { per_address_limit: u32 },
     MintTo { recipient: String },
-    MintFor { token_id: u64, recipient: String },
+    MintFor { token_id: u32, recipient: String },
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -41,7 +41,7 @@ pub enum QueryMsg {
 pub struct ConfigResponse {
     pub admin: Addr,
     pub base_token_uri: String,
-    pub num_tokens: u64,
+    pub num_tokens: u32,
     pub per_address_limit: u32,
     pub sg721_address: Addr,
     pub sg721_code_id: u64,
@@ -52,7 +52,7 @@ pub struct ConfigResponse {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct MintableNumTokensResponse {
-    pub count: u64,
+    pub count: u32,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
